moved commonly used function into parent, fleshed out shrink pixelator, added a run function to automatically execute color init, transform & write.
