+++ imaging/FloodFillPixelator.cpp
@@ -7,7 +7,6 @@ namespace imaging
 {
 	namespace
 	{
-		double colorDistance(PixelValue in_colorTo, PixelValue in_colorFrom);
 		int pixelToInt(const PixelValue& pixel);
 		PixelValue intToPixel(int pixelAsInt);
 	}
@@ -17,34 +16,6 @@ FloodFillPixelator::FloodFillPixelator(const std::string& in_path) : Pixelator(i
 bool FloodFillPixelator::to_pixels(int rows, int columns)
 {
 	return binarize() && pixelate(rows, columns);
-}
-
-bool FloodFillPixelator::binarize()
-{
-	ImgData hsvFile(pictureBuffer.cols, pictureBuffer.rows, pictureBuffer.type());
-	pictureBuffer.convertTo(hsvFile, cv::COLOR_BGR2HSV);
-	CV_Assert(hsvFile.channels() == 3);
-	ImgData outputFile = pictureBuffer.clone();
-	for (int i = 0; i < hsvFile.rows; ++i) {
-		for (int j = 0; j < hsvFile.cols; ++j) {
-			PixelValue outValue = resultColors[0];
-			PixelValue& content = hsvFile.at<cv::Vec3b>(i, j);
-			PixelValue& output = outputFile.at<cv::Vec3b>(i, j);
-			for (auto color : resultColors)
-			{
-				if (colorDistance(color, content) < colorDistance(outValue, content))
-				{
-					outValue = color;
-				}
-			}
-			for (auto channel = 0; channel < output.channels; ++channel)
-			{
-				output[channel] = outValue[channel];
-			}
-		}
-	}
-	pictureBuffer = outputFile.clone();
-	return true;
 }
 
 bool FloodFillPixelator::pixelate(int rows, int columns)
@@ -103,17 +74,9 @@ PixelValue FloodFillPixelator::getMostCommonColor(int start_row, int start_colum
 	}
 	return intToPixel(mostCommonColor);
 }
+
 namespace
 {
-	double colorDistance(PixelValue in_colorTo, PixelValue in_colorFrom)
-	{
-		uchar hueDiff = abs(in_colorTo[0] - in_colorFrom[0]);
-		uchar saturationDiff = abs(in_colorTo[1] - in_colorFrom[1]);
-		uchar valueDiff = abs(in_colorTo[2] - in_colorFrom[2]);
-		return hueDiff + 10. * (saturationDiff + 10. * valueDiff);
-	}
-
-
 	int pixelToInt(const PixelValue& pixel)
 	{
 		int pixelInt = 0;
