+++ imaging/Pixelator.h
@@ -10,15 +10,17 @@ namespace imaging
 	class Pixelator
 	{
 	public:
-		Pixelator(const std::string& in_path);
+		Pixelator(const std::string& in_path); 
+		virtual bool run(const std::string& in_writePath, const int rows, const int columns);
 		bool to_file(const std::string& in_path) const;
 		void add_color(PixelValue in_color);
 		bool empty() const;
-		int imgWidth() const;
-		int imgHeight() const;
 		virtual bool to_pixels(int rows, int columns) = 0;
 		virtual std::string imageStats() const;
+
 	protected:
+		bool binarize();
+
 		ImgData pictureBuffer;
 		std::vector<PixelValue> resultColors;
 	};
